requirements.txt
beautifulsoup4 
pandas
streamlit
chromadb
python-dotenv
openai
sentence-transformers
